+++ src/Core/Backend.Core/Services/IReadonlyService.cs
@@ -1,13 +1,14 @@
-﻿using System.Linq;
+﻿using System;
+using System.Linq;
 using System.Threading.Tasks;
 using GoldenEye.Shared.Core.Objects.DTO;
-using System;
 
-namespace GoldenEye.Backend.Core.Service
+namespace GoldenEye.Backend.Core.Services
 {
     public interface IReadonlyService<TDTO> : IDisposable where TDTO : class, IDTO
     {
         IQueryable<TDTO> Get();
+
         Task<TDTO> Get(int id);
     }
 }