+++ src/Core/Backend.Core/Services/IRestService.cs
@@ -1,13 +1,14 @@
 ﻿using System.Threading.Tasks;
 using GoldenEye.Shared.Core.Objects.DTO;
-using System;
 
-namespace GoldenEye.Backend.Core.Service
+namespace GoldenEye.Backend.Core.Services
 {
     public interface IRestService<TDTO> : IReadonlyService<TDTO> where TDTO : class, IDTO
     {
         Task<bool> Delete(int id);
+
         Task<TDTO> Post(TDTO dto);
+
         Task<TDTO> Put(TDTO dto);
     }
 }