+++ src/Core/Backend.Core.Tests/Changelog.md
@@ -1,1 +1,11 @@
-﻿
+﻿# v2.2.0 (27.12.2017)
+
+## Changes
+
+* Added [tests](Registration/RegistrationTests.cs) for [Registering all validators](../Backend.Core/Registration/Registration.cs) **[MAJOR]**
+
+# v1.0.0
+
+## Changes
+
+* Initial structure of project **[MAJOR]**